Add files via upload
Project: Student Web Portal in Php
-To download Student Web Portal project for free(Scroll Down)


Student Web Portal project is developed using PHP, CSS, HTML, and JavaScript. Talking about the project, it has all the essential features required for the management of web portal sites for the students. This project is mostly from the users perspective. The users has full access to the system. In this project, all the main functions are well validated.

About The System
The system is simple and quite easy to use. The basic fundamental of this system is to manage and upload the student profiles. Here the user can simply log in to the system and can use it. The user can upload his bio-data or CV or resume. From there you can easily search for jobs availability around your region. Also, you can simply upload your verified documents as well. This system allows you to create a chat system with those people who are using this system. As a whole you can refer this system as a web based job finding system.

To run this project you must have installed virtual server i.e XAMPP on your pc (for Windows). This Management System in PHP with source code is free to download, Use for educational purposes only! For the project demo, have a look at the image slider below.

https://www.youtube.com/watch?v=_Rb7gbJyl8A&t=49s
How To Run??

After Starting Apache and MySQL in XAMPP, follow the following steps

1st Step: Extract file
2nd Step: Copy the main project folder
3rd Step: Paste in xampp/htdocs/

Now Connecting Database

4th Step: Open a browser and go to URL “http://localhost/phpmyadmin/”
5th Step: Then, click on databases tab
6th Step: Create database naming “istudent_db” and then click on import tab
7th Step: Click on browse file and select “ istudent_db .sql” file
8th Step: Click on go.

9th Step: Open a browser and go to URL “http://localhost/student portal/”